+++ app/src/main/java/io/itch/fr/quizgame/feature_quiz/data/dao/QuestionDao.kt
@@ -1,0 +1,15 @@
+package io.itch.fr.quizgame.feature_quiz.data.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.Query
+import io.itch.fr.quizgame.feature_quiz.data.entities.Question
+
+@Dao
+interface QuestionDao {
+    @Query("SELECT * FROM questions")
+    suspend fun getAllQuestions(): List<Question>
+
+    @Insert
+    suspend fun insertQuestion(question: Question)
+}